Okay there's a lot to cover here, so see the desc. First of all, I've drastically changed the ob4 format. The functionality of the "nm" command has been split into starting a new mesh and changing the number of indices per face. The new mesh functionality has stayed in the "nm" command, and changing the number of indices has been moved to the "pt" command.
The "pt" command stands for "Primitive Type" and the
reason for that name will become clear the more you
read. Primitives are what I'm calling the "cells" of
the 4D shapes. For example, the pentachoron is
made up of 5 tetrahedra, and the hypercube is made
of 8 cubes: the tetrahedra and cubes being the
primitive types. Other examples of these primitive
types are octahedra, dodecahedra, icosahedra, and
in fact, any 3D polyhedra that can be decomposed
into tetrahedra can be primitive types... that is
if you can make an algorithm or have enough time to
manually figure out how to tetrahedralize them.
Tetrahedralizing is the same thing as triangulating,
except in one higher dimension.

The use of the "pt" command is simple, name which
primitive type you would like to use. For example,
the "pt" command for the pentachoron would be:
"pt Tetrahedron". Instead of saying which primitive
type to use, you may also use the keyword "Custom"
to make your own primitive. You may choose to give
the primitive a name so you can refer to it later
in the file (without having to redescribe it), or
you may leave it nameless. Nameless primitive types
will be overwritten when you switch to another
primitive type, so in a sense, they're temporary.

To describe the indices of a custom primitive,
there are 2 commands, "cm" and "ci" where "cc"
means "custom metadata" and "ci" means "custom
index". The "cm" command comes first, and only once.
--- As I'm writing this, I've realized I could've
--- put this in the pt command and not need an
--- extra command for this, but whatever.
Its job is to inform the parser how many tetrahedra
and how many vertices there are in the primitive.
Then the "ci" command contains the actual index
data. Each "ci" command describes one of the
tetrahedra that make up the primitive, so four
indices given that tetrahedra have four vertices.

Some things to consider:
1) You must set which primitive you want before
you include any face data.

--- Another "in the moment" thing: Tetrahedra
--- will be the default, but they aren't yet.

2) Types are used until another one is set.
3) Named custom types exist from where they are
created until the end of the file, whereas unnamed
types are temporary.
4) There is now error checking and I provide enough
information to be able to fix the issue.

Alright, after all that, there's still a little left.
One, aside from all 3D platonic solids being primitive
types, triangular and pentagonal prisms are as well;
rectangular prisms are taken care of as cubes, but they
don't really need to be cubes, just have 8 ordered
points and you're good.

Finally, some models have been added, and some have
been removed (but they're still in git if you really
want them for whatever reason).